Move order page to a popup
