+++ app/Base/components/Routes/index.tsx
@@ -29,12 +29,6 @@ function Routes(props: Props) {
                 </Route>
                 <Route
                     exact
-                    path={routes.orderPage.path}
-                >
-                    {routes.orderPage.load({ className })}
-                </Route>
-                <Route
-                    exact
                     path={routes.cartPage.path}
                 >
                     {routes.cartPage.load({ className })}
